Kalman filter skeleton
This code first initializes the noise covariance matrix for all the three sensors fused together (barometer, IMU, GPS) based on their spec sheets and then do the standard Kalman filter steps. However it is not very correct because the ideal method should be to use one of Barometer+IMU, GPS+IMU or barometer+GPS sensor fusion combination, depending on which two sensors are the most reliable. Unless it is feasible in other ways